feat(GenericContainer): introduce ExposedPortSetting trait for port management
Add the ExposedPortSetting trait to manage exposed ports in containers,
allowing ports to be set via a static variable or method. This change
removes redundant port management code from GenericContainer, enhancing
code reusability and maintainability. The new trait provides a more
flexible and consistent approach to defining and retrieving exposed
ports. Additionally, unit tests are added to ensure the functionality
of the ExposedPortSetting trait, verifying both static and dynamic port
exposure methods.
